Update 2023_Nakatsuka_CaliMexico.janno
I confronted the janno file with the supplementary table from Nakatsuka2023 and changed some stuff (see my review comment). There is a bunch of stuff that should probably be checked which I also mentioned in my comment but haven't changed here as I am not sure they actually need a change or they were intentionally altered as compared to the supplement.